Folder & file restructure to make things easier to understand.
