+++ components/landing.tsx
@@ -1,16 +1,14 @@
-import { Button, Center, Space, Text, Title } from '@mantine/core'
-import { IconArrowRight } from '@tabler/icons'
-import Link from 'next/link'
+import { Space, Title } from '@mantine/core'
 import React, { FunctionComponent } from 'react'
 import { useStyles } from '../styles/home'
-import { SponsorCarousel } from './carousel'
-import { DearMessages } from './dearMessages'
-import { Footer } from './footer'
-import { Media } from './media'
-import { NavBar } from './navbar'
-import { OurBlog } from './ourblog'
-import { OurSponsors } from './oursponsors'
-import { OurStory } from './ourstory'
+import { Footer } from './shared/footer'
+import { Media } from './landing/media'
+import { NavBar } from './shared/navbar'
+import { DearMessages } from './landing/dearMessages'
+import { SponsorCarousel } from './landing/carousel'
+import { OurSponsors } from './landing/oursponsors'
+import { OurStory } from './landing/ourstory'
+import { OurBlog } from './landing/ourblog'
 
 export const Landing: FunctionComponent = () => {
    const { classes } = useStyles()
